Move WKT_point_to_lonlat to geospatial.py
